Working on creating controllers

+++ K8s-hard-way/Certificates.sh
@@ -3,16 +3,22 @@
 # It is based on the tutorial from Kelsey Hightower: 
 # https://github.com/kelseyhightower/kubernetes-the-hard-way
 
+: '
+Options to use as variables:
+-m ==> eu-west-1-k8s-hardway-main ==> Master node name
+-w ==> k8s-hardway-nodes ==> Worker nodes name
+-l ==> k8s-hardway-nlb ===> Load balancer name
+'
 
 while getopts ":m:w:k:l:" opt; do
   case $opt in
     m) MASTER_NODE_NAME="$OPTARG";;
     w) WORKER_NODES_NAME="$OPTARG";;
-    k) KEY_FILE="$OPTARG";;
     l) LOADBALANCER_NAME="$OPTARG";;
     \?) echo "Invalid option -$OPTARG" >&2;;
   esac
 done
+
 
 checking_prerequisites(){
   # For this to work we need to install the following tools:
@@ -28,60 +34,73 @@ checking_prerequisites(){
     curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
     unzip awscliv2.zip
     sudo ./aws/install
-
-  else
     echo -e "\xE2\x9C\x94 aws-cli is installed"
-    echo "Checking if aws-cli is configured"
-
-    if ! [ -x "$(command -v aws configure)" ]; then
-      echo -e '\u2718 Error: aws-cli is not configured.' >&2
-      echo "Please run aws configure"
-      exit 1
-    fi
-
-    echo -e "\xE2\x9C\x94 aws-cli is configured"
-    echo "====Getting the key file===="
-
-    # expected output ===> i-0122b09d7f47e06f9
-    MASTER_NODE=$(aws ec2 describe-instances \
-      --filters "Name=tag:Name,Values=${MASTER_NODE_NAME}" \
-      --query "Reservations[*].Instances[*].InstanceId" \
-      --output text)
-
-    WORKER_NODES=$(aws ec2 describe-instances \
-      --filters "Name=tag:Name,Values=${WORKER_NODES_NAME}" \
-      --query "Reservations[*].Instances[*].InstanceId" \
-      --output text)
-
-    # expected output ===> k8s-hardway
-    KEY_FILE=$(aws ec2 describe-instances \
-      --filters "Name=instance-id,Values=${MASTER_NODE}" \
-      --query "Reservations[*].Instances[*].KeyName" \
-      --output text)
   fi
+
+  echo -e "\xE2\x9C\x94 aws-cli is installed"
+  echo "==== Checking if aws-cli is configured ===="
+
+  if ! [ -x "$(command -v aws configure)" ]; then
+    echo -e '\u2718 Error: aws-cli is not configured.' >&2
+    echo "Please run aws configure"
+    exit 1
+  fi
+
+  echo -e "\xE2\x9C\x94 aws-cli is configured"
+  echo "==== Getting the keys file ===="
+
+  # expected output ===> i-0122b09d7f47e06f9
+  MASTER_NODE=$(aws ec2 describe-instances \
+    --filters "Name=tag:Name,Values=${MASTER_NODE_NAME}" \
+    --query "Reservations[*].Instances[*].InstanceId" \
+    --output text)
+  echo -e "\xE2\x9C\x94 Master node is ${MASTER_NODE}"
+
+  WORKER_NODES=$(aws ec2 describe-instances \
+    --filters "Name=tag:Name,Values=${WORKER_NODES_NAME}" \
+    --query "Reservations[*].Instances[*].InstanceId" \
+    --output text)
+  echo -e "\xE2\x9C\x94 Worker nodes are ${WORKER_NODES}"
+
+  # expected output ===> k8s-hardway
+  KEY_FILE=$(aws ec2 describe-instances \
+    --filters "Name=instance-id,Values=${MASTER_NODE}" \
+    --query "Reservations[*].Instances[*].KeyName" \
+    --output text)
+  echo -e "\xE2\x9C\x94 Keys file is ${KEY_FILE}"
+
+  if [ ! -f "${KEY_FILE}.pem" ]; then
+    echo "\u2718 Error: ${KEY_FILE}.pem file does not exist" >&2
+    exit 1
+  fi
+  echo -e "\xE2\x9C\x94 ${KEY_FILE}.pem file exists"
+  
 
   # Check if cfssl and aws-cli are installed
   if ! [ -x "$(command -v cfssl)" ]; then
     echo '\u2718 Error: cfssl is not installed.' >&2
     echo "====Installing cfssl===="
     sudo apt-get install golang-cfssl
-  else
-    echo -e "\xE2\x9C\x94 cfssl is installed"
   fi
+  echo -e "\xE2\x9C\x94 cfssl is installed"
+  
 }
-# check if folder exists
 
+
+# check if folder exists
 change_dir(){
   if [ ! -d "certificates" ]; then
     mkdir certificates
   fi
+  echo "Changing directory to certificates"
+  cd certificates
 }
 
 define_certicate(){
   CN=$1
   O=$2
   FILE_NAME=$3
-  cat > certificates/${FILE_NAME}.json <<EOF
+  cat > ${FILE_NAME}.json <<EOF
 {
   "CN": "${CN}",
   "key": {
@@ -99,11 +118,13 @@ define_certicate(){
   ]
 }
 EOF
+
+  echo -e "\xE2\x9C\x94 ${FILE_NAME}.json is created"
 }
 
 # Genereate a CA certificate and private key
 generate_ca_certs(){
-  cat > certificates/ca-config.json <<EOF
+  cat > ca-config.json <<EOF
 {
   "signing": {
     "default": {
@@ -121,8 +142,13 @@ EOF
   FILE_NAME="ca-csr"
   define_certicate "Kubernetes" "Kubernetes" ${FILE_NAME}
   
-  cfssl gencert -initca certificates\ca-csr.json | cfssljson -bare ca
+  cfssl gencert -initca ${FILE_NAME}.json | cfssljson -bare ca
+
+  # Check on the certificates
+  openssl x509 -in ca.pem -text -noout
+
 }
+
 
 # CA certificate for the admin controller
 admin_controller(){
@@ -134,12 +160,13 @@ admin_controller(){
     -ca-key=ca-key.pem \
     -config=ca-config.json \
     -profile=kubernetes \
-    certificate/${FILE_NAME}.json | cfssljson -bare admin
+    ${FILE_NAME}.json | cfssljson -bare admin
 }
+
 
 # Generate the CA for the worker nodes
 worker_nodes(){
-  for instance in worker-0 worker-1 worker-2; do
+  for instance in $WORKER_NODES; do
 
     FILE_NAME="${instance}-csr"
     define_certicate "system:node:${instance}" "system:nodes" ${FILE_NAME}
@@ -160,9 +187,10 @@ worker_nodes(){
       -config=ca-config.json \
       -hostname=${instance},${EXTERNAL_IP},${INTERNAL_IP} \
       -profile=kubernetes \
-      certificate/${FILE_NAME}.json | cfssljson -bare ${instance}
+      ${FILE_NAME}.json | cfssljson -bare ${instance}
   done
 }
+
 
 # certificate for the control manager
 control_manger(){
@@ -175,8 +203,9 @@ control_manger(){
     -ca-key=ca-key.pem \
     -config=ca-config.json \
     -profile=kubernetes \
-    certificate/${FILE_NAME}.json | cfssljson -bare kube-controller-manager
+    ${FILE_NAME}.json | cfssljson -bare kube-controller-manager
 }
+
 
 # certificate for the kube-proxy
 kube_proxy(){
@@ -188,8 +217,9 @@ kube_proxy(){
     -ca-key=ca-key.pem \
     -config=ca-config.json \
     -profile=kubernetes \
-    certificate/${FILE_NAME}.json | cfssljson -bare kube-proxy
+    ${FILE_NAME}.json | cfssljson -bare kube-proxy
 }
+
 
 # certificate for the scheduler
 scheduler(){
@@ -201,34 +231,38 @@ scheduler(){
     -ca-key=ca-key.pem \
     -config=ca-config.json \
     -profile=kubernetes \
-    certificate/${FILE_NAME}.json | cfssljson -bare kube-scheduler
+    ${FILE_NAME}.json | cfssljson -bare kube-scheduler
 }
 
+
 # certificate for the Kubernetes API server
-: '
-For this to work we will need a load balancer in front of the Kubernetes API servers
-The load balancer will have a static IP address assigned to it
-'
 api_server(){
   FILE_NAME="kubernetes-csr"
   define_certicate "kubernetes" "Kubernetes" ${FILE_NAME}
 
   ## Get the public address of the load balancer
-  KUBERNETES_PUBLIC_ADDRESS=$(aws elbv2 describe-load-balancers\
-  --names ${LOADBALANCER_NAME}\
-  --query "LoadBalancers[].AvailabilityZones[].LoadBalancerAddresses[].IpAddress[]"\
-  --output text)
+  MASTER_PUBLIC_IP=$(aws ec2 describe-instances \
+        --filters "Name=instance-id,Values=${MASTER_NODE}" \
+        --query "Reservations[*].Instances[*].PublicIpAddress" \
+        --output text)
 
   KUBERNETES_HOSTNAMES=kubernetes,kubernetes.default,kubernetes.default.svc,kubernetes.default.svc.cluster,kubernetes.svc.cluster.local
 
+  local INTERNAL_IP=$(aws ec2 describe-instances \
+        --filters "Name=instance-id,Values=${MASTER_NODE}" \
+        --query "Reservations[*].Instances[*].PrivateIpAddress" \
+        --output text)
+
+  # Check on the ips used
   cfssl gencert \
     -ca=ca.pem \
     -ca-key=ca-key.pem \
     -config=ca-config.json \
-    -hostname=10.32.0.1,10.240.0.10,10.240.0.11,10.240.0.12,${KUBERNETES_PUBLIC_ADDRESS},127.0.0.1,${KUBERNETES_HOSTNAMES} \
+    -hostname=${INTERNAL_IP},${MASTER_PUBLIC_IP},127.0.0.1,${KUBERNETES_HOSTNAMES} \
     -profile=kubernetes \
-    certificate/${FILE_NAME}.json | cfssljson -bare kubernetes
+    ${FILE_NAME}.json | cfssljson -bare kubernetes
 }
+
 
 # certificate for the Kubernetes Service Account
 service_account(){
@@ -240,25 +274,33 @@ service_account(){
     -ca-key=ca-key.pem \
     -config=ca-config.json \
     -profile=kubernetes \
-    certificate/${FILE_NAME}.json | cfssljson -bare service-account
+    ${FILE_NAME}.json | cfssljson -bare service-account
 }
 
-# Distribute the certificates and keys to the worker nodes
-distribute_workers{
-  for worker_instance in $WORKER_NODES; do
 
+# Distribute the certificates and keys to the worker nodes
+distribute_workers(){
+
+  for worker_instance in $WORKER_NODES; do
     # result should  be like this ec2-34-254-19-53.eu-west-1.compute.amazonaws.com
     DNS_NAME=$(aws ec2 describe-instances \
     --filters "Name=instance-id,Values=${worker_instance}" \
     --query "Reservations[*].Instances[*].PublicDnsName" \
     --output text)
+    
+    echo -e "\xE2\x9C\x94 ${worker_instance} DNS name is ${DNS_NAME}"
 
-    scp -i ${KEY_FILE}.pem  ${instance}-key.pem ${instance}.pem ubuntu@${DNS_NAME}:~/
+    sudo scp -i ../${KEY_FILE}.pem ${worker_instance}-key.pem ${worker_instance}.pem ca.pem ubuntu@${DNS_NAME}:~/
+
+    # success message
+    echo -e "\xE2\x9C\x94 ${worker_instance} files have been distributed successfully"
   done
 }
 
+
 # Distribute the certificates and keys to the controller nodes
 distribute_controllers(){
+  # check if key file exists
   for master_instance in $MASTER_NODE; do
 
     DNS_NAME=$(aws ec2 describe-instances \
@@ -266,21 +308,26 @@ distribute_controllers(){
     --query "Reservations[*].Instances[*].PublicDnsName" \
     --output text)
 
-    scp -i ${KEY_FILE} ca.pem ca-key.pem kubernetes-key.pem kubernetes.pem \
+    echo -e "\xE2\x9C\x94 ${master_instance} DNS name is ${DNS_NAME}"
+
+    sudo scp -i ../${KEY_FILE}.pem ca.pem ca-key.pem kubernetes-key.pem kubernetes.pem \
       service-account-key.pem service-account.pem ubuntu@${DNS_NAME}:~/
+    
+    # success message
+    echo -e "\xE2\x9C\x94 ${instance} files have been distributed successfully"
   done
 }
 
 
 checking_prerequisites
 change_dir
-generate_ca_certs
-admin_controller
-worker_nodes
-control_manger
-kube_proxy
-scheduler
-api_server
-service_account
-distribute_workers
+# generate_ca_certs
+# admin_controller
+# worker_nodes
+# control_manger
+# kube_proxy
+# scheduler
+# api_server
+# service_account
+# distribute_workers
 distribute_controllers
